requirements.txt
fastapi
pydantic
aiodocker
uvicorn
python-dotenv